pyproject.toml
[tool.poetry]
name = "logistic-regression"
version = "0.1.0"
description = "Logistic-regression for Social_Network_ads problem"
authors = ["Pavel Perestoronin<true.perestoronin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.26.2"
scikit-learn = "^1.3.2"
pandas = "^2.1.4"
hydra-core = "^1.3.2"
pre-commit = "^3.6.0"
mlflow = "^2.9.2"
flake8 = "^6.1.0"
dvc = "^3.36.1"
dvc-gdrive = "^3.0.0"
tensorboard = "^2.15.1"
torch = "^2.1.2"
black = "^23.12.1"
isort = "^5.13.2"
loguru = "^0.7.2"
skops = "^0.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
